# openapi.yaml
openapi: 3.0.0
info:
  title: LeanStock API
  version: 1.0.0
  description: Inventory management system with multi-tenancy, transaction safety and race condition protection
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        example: 1
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        example: 10
  responses:
    BadRequest:
      description: Invalid request
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Resource not found
    Conflict:
      description: Conflict occurred
...
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: 1
        name:
          type: string
          example: "Milk"
        price:
          type: number
          example: 5000
...
    Inventory:
      type: object
      properties:
        productId:
          type: integer
          example: 1
        shopId:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 10
    StockTransferRequest:
      type: object
      required: [productId, fromShopId, toShopId, quantity]
      properties:
        productId:
          type: integer
          example: 1
        fromShopId:
          type: integer
          description: Source shop ID (must belong to your tenant)
          example: 1
        toShopId:
          type: integer
          description: Destination shop ID (must belong to your tenant)
          example: 2
        quantity:
          type: integer
          minimum: 1
          example: 5
    StockTransferResponse:
      type: object
      properties:
        message:
          type: string
          example: Transfer successful
        productId:
          type: integer
          example: 1
        fromShopId:
          type: integer
          example: 1
        toShopId:
          type: integer
          example: 2
        quantity:
          type: integer
          example: 5
    InventoryCreateRequest:
      type: object
      required: [productId, shopId, quantity]
      properties:
        productId:
          type: integer
          example: 1
        shopId:
          type: integer
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 20
...
paths:
...
  /products:
    get:
      summary: Get products
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    type: object
                    properties:
                      page: { type: integer, example: 1 }
                      limit: { type: integer, example: 10 }
                      total: { type: integer, example: 42 }
                      totalPages: { type: integer, example: 5 }
                      hasNext: { type: boolean }
                      hasPrev: { type: boolean }
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create product
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        "201":
          description: Product created
        "400":
          $ref: '#/components/responses/BadRequest'
        "409":
          $ref: '#/components/responses/Conflict'
  /products/{id}:
    get:
      summary: Get product by ID
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "404":
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number

      responses:
        "200":
          description: Updated
        "400":
          description: Bad request
        "404":
          description: Product not found
        "500":
          description: Server error

    delete:
      summary: Delete product
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders:
    get:
      summary: Get orders
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        "200":
          description: Paginated orders list
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shopId, items]
              properties:
                shopId:
                  type: integer
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: integer
                      quantity:
                        type: integer
      responses:
        "201":
          description: Order created
        "400":
          $ref: '#/components/responses/BadRequest'
  /orders/{id}:
    get:
      summary: Get order by ID
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
        "404":
          $ref: '#/components/responses/NotFound'

  # Короткий алиас (то же поведение, что POST /inventory/transfer)
  /transfer:
    post:
      summary: Transfer stock between shops (alias)
      description: |
        Идентично `POST /inventory/transfer`. MANAGER, Bearer JWT, ACID-транзакция.
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockTransferRequest'
            examples:
              default:
                summary: Move five units shop 1 → 2
                value:
                  productId: 1
                  fromShopId: 1
                  toShopId: 2
                  quantity: 5
      responses:
        "200":
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockTransferResponse'
        "400":
          description: Missing or invalid fields
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Forbidden — requires MANAGER role
        "404":
          description: Product or shop not found for your tenant
        "409":
          description: Insufficient stock at source shop
  /inventory:
    get:
      summary: Get inventory for your tenant
      description: |
        Returns inventory rows only for shops that belong to the authenticated user's tenant.
        Roles: any authenticated user (Bearer JWT).
      tags: [Inventory]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        "200":
          description: Paginated inventory list with product and shop
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inventory'
                  pagination:
                    type: object
        "401":
          $ref: '#/components/responses/Unauthorized'
  /inventory/transfer:
    post:
      summary: Transfer stock between shops
      description: |
        Moves `quantity` units of `productId` from `fromShopId` to `toShopId`.
        Runs in a single ACID transaction. Source must have sufficient stock.

        **Authorization:** Bearer JWT — role **MANAGER** only.

        Both shops must belong to the same tenant as the JWT. Emails queued (async).
        Writes an audit log entry `STOCK_TRANSFER`.
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockTransferRequest'
            examples:
              default:
                summary: Move five units shop 1 → 2
                value:
                  productId: 1
                  fromShopId: 1
                  toShopId: 2
                  quantity: 5
      responses:
        "200":
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockTransferResponse'
              example:
                message: Transfer successful
        "400":
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "productId, fromShopId, toShopId, and quantity are required"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Forbidden — requires MANAGER role
        "404":
          description: Product not found or shop not in your tenant
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: "Product not found"
        "409":
          description: Insufficient stock at source shop
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Insufficient stock to transfer
  /inventory/add:
    post:
      summary: Add inventory line (stock record)
      description: |
        Creates inventory for a `(productId, shopId)` pair. Fails if the pair already exists (409).

        **Authorization:** Bearer JWT — role **MANAGER**. Shop must be in your tenant.
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryCreateRequest'
      responses:
        "201":
          description: Inventory row created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        "400":
          description: Missing fields
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Forbidden — requires MANAGER
        "404":
          description: Shop not found
        "409":
          description: Inventory already exists for this product and shop
...